Deploying Agents to AWS with Bedrock AgentCore/Developing AgentCore Apps Locally/requirements.txt
boto3==1.40.3
botocore==1.40.3
bedrock-agentcore==0.1.1
strands-agents==1.4.0
strands-agents-tools==0.2.3